requirements.txt
numpy
pandas
Pillow
face-recognition
opencv-python
mysql-connector-python